arrange : move to Person Package about Person
