+++ src/test/kotlin/study/person/Person.kt
@@ -1,3 +1,3 @@
-package study
+package study.person
 
 data class Person(val name: String, val age: Int, var nickname: String? = name)
