+++ src/test/kotlin/study/person/PersonTest.kt
@@ -1,4 +1,4 @@
-package study
+package study.person
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
